Fleshed out Library class definition function blocks and comments
I decided to comment the hell of the functions in the Library class since
they're actually kind of terrifying. Not really following any standard I
know of, but it should get the point across. Each comment block contains
the Function name, a brief description of what it's for, what arguments it
takes, and what information it returns.