Replaced getAllEntitiesWithComponent() with allEntityIDs()
It makes more sense to me, currently, to simply provide systems with a
list of all the available Entities rather than a list containing Entities
with a single Component type. If there was a better way to search for
multiple Component types and return a list of those, however, I'd include
a function for that as well.