Created definitions for World class
There's a few edits made to the header for this class as well. It was
only after I started crafting the function definitions that I realized a
few passed arguments were incorrect; namely I changed the argument in
addSystem() to a pass-by-reference, and updated the systems vector to
store pointers to Systems instead of Systems themselves.